Air Current
Created a rudimentary air current that detects when the character comes into contact with the air current object. While the player is colliding with the object they gradually rise. NOTE: need to work the currents rising speed as it is dependent on when the player collides with the air current. This is most likely due to the use of gravity.